// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PessoaStatus = {
  EM_DIA: 'EM_DIA',
  PRE_ASSOCIADO: 'PRE_ASSOCIADO',
  DOCUMENTACAO_OK: 'DOCUMENTACAO_OK',
  ATENCAO: 'ATENCAO',
  ISENTO: 'ISENTO',
  SAIU: 'SAIU'
} as const

export type PessoaStatus = (typeof PessoaStatus)[keyof typeof PessoaStatus]


export const PessoaTipo = {
  HUMANO: 'HUMANO',
  PET: 'PET'
} as const

export type PessoaTipo = (typeof PessoaTipo)[keyof typeof PessoaTipo]


export const PessoaSexo = {
  M: 'M',
  F: 'F',
  OUTRO: 'OUTRO'
} as const

export type PessoaSexo = (typeof PessoaSexo)[keyof typeof PessoaSexo]


export const DocumentoTipo = {
  FICHA_FILIACAO: 'FICHA_FILIACAO',
  RG_CPF: 'RG_CPF',
  COMP_RESIDENCIA: 'COMP_RESIDENCIA',
  RECEITA_MEDICA: 'RECEITA_MEDICA',
  LAUDO: 'LAUDO'
} as const

export type DocumentoTipo = (typeof DocumentoTipo)[keyof typeof DocumentoTipo]


export const ProdutoTipo = {
  OLEO: 'OLEO',
  POMADA: 'POMADA',
  RESINA: 'RESINA',
  RESGATE: 'RESGATE',
  SUPOSITORIO: 'SUPOSITORIO',
  FLORES: 'FLORES'
} as const

export type ProdutoTipo = (typeof ProdutoTipo)[keyof typeof ProdutoTipo]


export const OleoPerfil = {
  FULL_SPECTRUM: 'FULL_SPECTRUM',
  BROAD_SPECTRUM: 'BROAD_SPECTRUM',
  ISOLADO: 'ISOLADO'
} as const

export type OleoPerfil = (typeof OleoPerfil)[keyof typeof OleoPerfil]


export const OleoVolume = {
  ML_10: 'ML_10',
  ML_30: 'ML_30'
} as const

export type OleoVolume = (typeof OleoVolume)[keyof typeof OleoVolume]